Hazelcast_speed_test
This code is used to create Hazelcast instance and measure Hazelcast's storage performance. The code creates a Hazelcast instance named hazelcastInstance and obtains a datamap named "randomData". Later,
it uses a Random object to generate random numbers from the data and a for loop to insert the numbers into the datamap. Before each data is added, the start time is measured, and after all the data is added, the end time is measured. The difference between these two times measures the time it takes for data to be added to the data map.

Same thing,
is also done for fetching the data from the data map. First the start time is measured, the data is taken from the data map, and then the end time is measured. The difference between these two times measures the time it takes for data to be retrieved from the data map. The results are printed to the screen.