Basic calculator with bash script
This script performs a simple arithmetic calculation based on the input from the user. The user is first prompted to input the calculation operand, which can be one of the following: +, -, *, /, %, or !.

If the operand is not "!", the user will be prompted to input two numbers, "number1" and "number2". If the operand is "!", the user will only be prompted to input one number.

Based on the input operand, the script performs the corresponding arithmetic calculation using the input numbers, and the result is stored in the "result" variable.

The final step is to print the result to the screen. If the input operand is invalid, an error message is displayed, and the script exits with a status code of 1.